fix bug 1456956 - trigger off of processed crashes
This redoes the submitter to trigger off of processed crashes instead of
raw crashes. The submitter's mission is to submit some percent of the
crashes that *were processed* to another environment. Given that, it
makes more sense to trigger off of saved processed crashes and not
trigger off of saved raw crashes and check the throttle instruction.