bug 1665365: fix logging issues
This tweaks the code to use the same logger name everywhere. I don't
think this affects things, but maybe it does.

This adds an explicit import for dockerflow at the top. If it uses
things that aren't available, we'll get a better import error. Maybe the
error is getting swallowed up in logging configuration.