bug 1757612: remove collector-added keys
The socorro-submitter listens for data saved by the prod collector,
fetches that data, and then packages it up as a crash report to submit
to the stage environment.

We don't want the submitter to package up keys the collector added so
this removes those keys from the payload before packaging the crash
report up.

While doing that, I also fixed some comments that were wrong.