bug 1687987: support version 2 raw crash
We're adjusting the structure of the raw crash so that
collector-generated fields are primarily in "metadata". That makes it
easier for a lot of things including the submitter since it allows us to
add new fields to the collector without having to do a followup fix to
the submitter to remove the keys from the crash report before submitting
to stage.

This also adds a bunch of tests.