bug-1895817: add support for multiple destinations
This adds support for multiple destinations by using a new
SUBMITTER_DESTINATIONS configuration variable.

This deprecates the old SUBMITTER_DESTINATION_URL and SUBMITTER_THROTTLE
variables but provides support for both ways of configuring the
submitter with SUBMITTER_DESTINATIONS taking precedence.

This adds tests for the deprecated variables as well as multiple
destinations and differing throttles.

In order to implement this, this rearranges the handler function, but
the work the handler function does should be the same modulo posting to
multiple destinations.

Note: This doesn't change the ``Throttleable`` header, so the
destination collector will throttle incoming crash reports from the
socorro-submitter.